+++ src/pages/contractPage/components/ProductDetail.tsx
